Very solid progress on multi-selection menu

+++ Flashcards.Wolfieeex/Flashcards.Wolfieeex/View/UserInterface/FlashcardMenu.cs
@@ -4,6 +4,7 @@ using Spectre.Console;
 using System.ComponentModel;
 using System.Text.RegularExpressions;
 using static Flashcards.Wolfieeex.Model.InputValidationEnums;
+using static Flashcards.Wolfieeex.Model.MultiInputMenuEnums;
 using static Flashcards.Wolfieeex.Model.SelectionEnums;
 
 namespace Flashcards.Wolfieeex.View.UserInterface;
@@ -12,9 +13,13 @@ internal class FlashcardMenu : Menu
 {
 	public FlashcardMenu() : base(Color.Orange4) { }
 
-	protected override void DisplayMenu()
+	public override void DisplayMenu()
 	{
 		Console.Clear();
+
+		var menuSelections = Enum.GetValues(typeof(FlashcardChoices)).Cast<FlashcardChoices>();
+		
+		string[] choices = { "1", "2", "3" };
 
 		bool IsMenuRunning = true;
 		while (IsMenuRunning)
@@ -22,12 +27,7 @@ internal class FlashcardMenu : Menu
 			var userChoice = AnsiConsole.Prompt(
 				new SelectionPrompt<FlashcardChoices>()
 				.Title("What would you like to do today?")
-				.AddChoices(
-					FlashcardChoices.ViewFlashcards,
-					FlashcardChoices.AddFlashcard,
-					FlashcardChoices.UpdateFlashcard,
-					FlashcardChoices.DeleteFlashcard,
-					FlashcardChoices.ReturnToMainMenu)
+				.AddChoices(menuSelections)
 				.UseConverter(s => GetDisplayName(s))
 				.HighlightStyle(style)
 				.WrapAround()
@@ -80,23 +80,34 @@ internal class FlashcardMenu : Menu
 	private void UpdateFlashcard()
 	{
 		var stackId = ChooseStack("Choose the stack where the flashcard is: ");
-		var flashcardId = ChooseFlashcard("Choose flashcard to update", stackId);
+
+		/*var flashcardId = ChooseFlashcard("Choose flashcard to update", stackId);
 
 		var propertiesToUpdate = new Dictionary<string, object>();
 
-		//if (Ansi)
+		if (Ansi)*/
 	}
 
 	private void DeleteFlashcard()
 	{
-		var stackId = ChooseStack("Where is the flashcard you want to delete?:");
-		var flashcard = ChooseFlashcard("Choose flashcard you want to delete", stackId);
+		bool menuWhile = true;
+		while (menuWhile)
+		{
+			var stackId = ChooseStack("Where is the flashcard you want to delete?:");
+			if (stackId == -1)
+				return;
 
-		if (!AnsiConsole.Confirm("Are you sure?"))
-			return;
+			var flashcard = ChooseFlashcard("Choose flashcard you want to delete", stackId);
+			if (flashcard == -1)
+				continue;
 
-		var dataAccess = new DataAccess();
-		dataAccess.DeleteFlashcard(flashcard);
+			if (!AnsiConsole.Confirm("Are you sure?"))
+				continue;
+
+			var dataAccess = new DataAccess();
+			dataAccess.DeleteFlashcard(flashcard);
+			menuWhile = false;
+		}
 	}
 
 	private void AddFlashcard()
@@ -104,10 +115,14 @@ internal class FlashcardMenu : Menu
 		Flashcard flashcard = new();
 
 		string dummyInput = "";
-		flashcard.StackId = ChooseStack("Choose one of your previous stacks:");
+
+		AddFlashcardMenu addFlashcardMenu = new(menuColors.UserInputColor);
+		addFlashcardMenu.DisplayMenu();	
+
+		/*flashcard.StackId = ChooseStack("Choose one of your previous stacks:");
 
 		flashcard.Question = GetQuestion();
-		flashcard.Answer = GetAnswer();
+		flashcard.Answer = GetAnswer();*/
 
 		var dataAccess = new DataAccess();
 		dataAccess.InsertFlashcard(flashcard);
@@ -121,9 +136,10 @@ internal class FlashcardMenu : Menu
 		var stacksArray = stacks.Select(x => x.Name).ToArray();
 		var option = AnsiConsole.Prompt(new SelectionPrompt<string>()
 			.Title(message)
+			.AddChoices("[grey]Return to previous menu[/]")
 			.AddChoices(stacksArray));
 
-		return stacks.Single(x => x.Name == option).Id;
+		return stacks.Single(x => x.Name == option)?.Id ?? -1;
 	}
 
 	/// <returns>Returns -1 if user returns to previous menu without selection.</returns>
@@ -135,7 +151,10 @@ internal class FlashcardMenu : Menu
 		var flashcardsArray = flashcards.Select(x => x.Question).ToArray();
 		var option = AnsiConsole.Prompt(new SelectionPrompt<string>()
 			.Title(message)
-			.AddChoices(flashcardsArray));
+			.AddChoices("[grey]Return to previous menu[/]")
+			.AddChoices(flashcardsArray)
+			);
+			
 
 		return flashcards.Single(x => x.Question == option)?.Id ?? -1;
 	}
